feat: decouple code component
